refactor: rename file and guild title -> name
